+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_05_17_021939) do
+ActiveRecord::Schema[7.0].define(version: 2022_05_21_051327) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "citext"
   enable_extension "plpgsql"
@@ -131,6 +131,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_05_17_021939) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.string "middle_name"
+    t.string "phone"
     t.index ["last_name"], name: "index_residents_on_last_name"
   end